Moved RepositoryViewModel to the repository package. (now for sure :P)
