+++ app/src/main/java/com/abishov/hexocat/home/trending/TrendingFragment.java
@@ -19,6 +19,7 @@ import com.abishov.hexocat.Hexocat;
 import com.abishov.hexocat.R;
 import com.abishov.hexocat.commons.views.BaseFragment;
 import com.abishov.hexocat.home.repository.RepositoryAdapter;
+import com.abishov.hexocat.home.repository.RepositoryViewModel;
 import com.jakewharton.rxbinding2.view.RxView;
 import com.squareup.picasso.Picasso;
