+++ app/src/test/java/com/abishov/hexocat/home/trending/RepositoryViewModelUnitTests.java
@@ -1,5 +1,7 @@
 package com.abishov.hexocat.home.trending;
 
+import com.abishov.hexocat.home.repository.RepositoryViewModel;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
